# docs/swagger.yaml
basePath: /api
definitions:
  APIError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  APIErrors:
    properties:
      code:
        type: integer
      errors:
        items:
          $ref: '#/definitions/APIError'
        type: array
    type: object
...
  CreateProductRequest:
    properties:
      ean:
        type: string
      height:
        type: number
      image:
        type: string
      is_enable:
        type: boolean
      isbn:
        type: string
      jan:
        type: string
      length:
        type: number
      location:
        type: string
      manufacturer_id:
        type: string
      media_ids:
        items:
          type: string
        type: array
      minimum:
        type: integer
      mpn:
        type: string
      price_business:
        type: number
      price_retail:
        type: number
      price_wholesale:
        type: number
      quantity:
        type: integer
      sku:
        type: string
      sort_order:
        type: integer
      stock_status:
        type: string
      subtract:
        type: boolean
      upc:
        type: string
      weight:
        type: number
      width:
        type: number
    required:
    - minimum
    type: object
...
  FullPagingData:
    properties:
      last_page:
        type: integer
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  ImageDTO:
    properties:
      alt:
        type: string
      height:
        type: integer
      id:
        type: string
      presets:
        additionalProperties:
          additionalProperties:
            type: string
          type: object
        type: object
      width:
        type: integer
    type: object
...
  JSONResponse-ProductResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ProductResponse'
      message:
        type: string
    type: object
  JSONResponse-ProductReviewResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ProductReviewResponse'
      message:
        type: string
    type: object
...
  JSONResponse-ProductWithMediumResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ProductWithMediumResponse'
      message:
        type: string
    type: object
...
  JSONResponse-ResponseWithFullPagination-ProductResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ResponseWithFullPagination-ProductResponse'
      message:
        type: string
    type: object
  JSONResponse-ResponseWithFullPagination-ProductReviewResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ResponseWithFullPagination-ProductReviewResponse'
      message:
        type: string
    type: object
...
  JSONResponse-ResponseWithFullPagination-github_com_stickpro_go-store_internal_dto_EnrichedVariantDTO:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/ResponseWithFullPagination-github_com_stickpro_go-store_internal_dto_EnrichedVariantDTO'
      message:
        type: string
    type: object
...
  JSONResponse-VariantListResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/VariantListResponse'
      message:
        type: string
    type: object
...
  ProductResponse:
    properties:
      ean:
        type: string
      height:
        type: number
      id:
        type: string
      image:
        type: string
      is_enable:
        type: boolean
      isbn:
        type: string
      jan:
        type: string
      length:
        type: number
      location:
        type: string
      manufacturer_id:
        $ref: '#/definitions/uuid.NullUUID'
      minimum:
        type: integer
      mpn:
        type: string
      price_business:
        type: number
      price_retail:
        type: number
      price_wholesale:
        type: number
      quantity:
        type: integer
      sku:
        type: string
      sort_order:
        type: integer
      stock_status:
        type: string
      subtract:
        type: boolean
      upc:
        type: string
      variant:
        $ref: '#/definitions/ProductVariantResponse'
      weight:
        type: number
      width:
        type: number
    type: object
  ProductReviewResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  ProductVariantResponse:
    properties:
      category_id:
        $ref: '#/definitions/uuid.NullUUID'
      description:
        type: string
      id:
        type: string
      is_enable:
        type: boolean
      meta_description:
        type: string
      meta_h1:
        type: string
      meta_keyword:
        type: string
      meta_title:
        type: string
      model:
        type: string
      name:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
    type: object
  ProductWithMediumResponse:
    properties:
      images:
        items:
          $ref: '#/definitions/github_com_stickpro_go-store_internal_dto.ImageDTO'
        type: array
      product:
        $ref: '#/definitions/ProductResponse'
    type: object
...
  ResponseWithFullPagination-ProductResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/ProductResponse'
        type: array
      pagination:
        $ref: '#/definitions/FullPagingData'
    type: object
  ResponseWithFullPagination-ProductReviewResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/ProductReviewResponse'
        type: array
      pagination:
        $ref: '#/definitions/FullPagingData'
    type: object
...
  ResponseWithFullPagination-github_com_stickpro_go-store_internal_dto_EnrichedVariantDTO:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_stickpro_go-store_internal_dto.EnrichedVariantDTO'
        type: array
      pagination:
        $ref: '#/definitions/FullPagingData'
    type: object
...
  StockStatus:
    enum:
    - IN_STOCK
    - PRE_ORDER
    - OUT_OF_STOCK
    type: string
    x-enum-varnames:
    - InStock
    - PreOrder
    - OutOfStock
...
  UpdateProductRequest:
    properties:
      ean:
        type: string
      height:
        type: number
      is_enable:
        type: boolean
      isbn:
        type: string
      jan:
        type: string
      length:
        type: number
      location:
        type: string
      manufacturer_id:
        type: string
      media_ids:
        items:
          type: string
        type: array
      minimum:
        type: integer
      model:
        type: string
      mpn:
        type: string
      price_business:
        type: number
      price_retail:
        type: number
      price_wholesale:
        type: number
      quantity:
        type: integer
      sku:
        type: string
      sort_order:
        type: integer
      stock_status:
        type: string
      subtract:
        type: boolean
      upc:
        type: string
      weight:
        type: number
      width:
        type: number
    type: object
...
  VariantCardResponse:
    properties:
      category_id:
        $ref: '#/definitions/uuid.NullUUID'
      description:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/ImageDTO'
      is_enable:
        type: boolean
      manufacturer_id:
        $ref: '#/definitions/uuid.NullUUID'
      model:
        type: string
      name:
        type: string
      price_business:
        type: number
      price_retail:
        type: number
      price_wholesale:
        type: number
      product_id:
        type: string
      slug:
        type: string
      stock_status:
        type: string
    type: object
...
  VariantListResponse:
    properties:
      facet_stats:
        additionalProperties:
          $ref: '#/definitions/github_com_stickpro_go-store_internal_dto.CategoryFacetStat'
        type: object
      facets:
        additionalProperties:
          additionalProperties:
            format: int64
            type: integer
          type: object
        type: object
      items:
        items:
          $ref: '#/definitions/VariantCardResponse'
        type: array
      pagination:
        $ref: '#/definitions/FullPagingData'
    type: object
...
  github_com_stickpro_go-store_internal_delivery_http_request_product_review_request.CreateProductReviewRequest:
    properties:
      body:
        type: string
      rating:
        type: integer
      title:
        type: string
      variant_id:
        type: string
    type: object
...
  github_com_stickpro_go-store_internal_dto.CategoryFacetStat:
    properties:
      max:
        type: number
      min:
        type: number
    type: object
  github_com_stickpro_go-store_internal_dto.EnrichedVariantDTO:
    properties:
      category_id:
        $ref: '#/definitions/uuid.NullUUID'
      category_ids:
        description: |-
          CategoryIDs holds the variant's category plus all its ancestors (and the same
          for its additional categories) so the search index can filter a whole subtree.
        items:
          type: string
        type: array
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      description:
        $ref: '#/definitions/pgtype.Text'
      id:
        type: string
      image:
        allOf:
        - $ref: '#/definitions/github_com_stickpro_go-store_internal_dto.ImageDTO'
        description: |-
          Image is the product's main image (first by product_media.sort_order), baked into
          the search document; nil when the product has no gallery.
      is_enable:
        type: boolean
      manufacturer_id:
        $ref: '#/definitions/uuid.NullUUID'
      meta_description:
        $ref: '#/definitions/pgtype.Text'
      meta_h1:
        $ref: '#/definitions/pgtype.Text'
      meta_keyword:
        $ref: '#/definitions/pgtype.Text'
      meta_title:
        $ref: '#/definitions/pgtype.Text'
      model:
        type: string
      name:
        type: string
      price_business:
        type: number
      price_retail:
        type: number
      price_wholesale:
        type: number
      product_id:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
      stock_status:
        $ref: '#/definitions/StockStatus'
      updated_at:
        $ref: '#/definitions/pgtype.Timestamp'
      viewed:
        type: integer
    type: object
  github_com_stickpro_go-store_internal_dto.ImageDTO:
    properties:
      alt:
        type: string
      height:
        type: integer
      id:
        type: string
      presets:
        additionalProperties:
          additionalProperties:
            type: string
          type: object
        type: object
      width:
        type: integer
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
  uuid.NullUUID:
    properties:
      uuid:
        type: string
      valid:
        description: Valid is true if UUID is not NULL
        type: boolean
    type: object
info:
  contact:
    email: go-store@stick.sh
    name: Vladislav B
  description: This is an API for go-store
  title: GO-store
  version: "1.0"
paths:
...
  /api/v1/product_review/:
    post:
      consumes:
      - application/json
      description: Create Product Review
      parameters:
      - description: Create Product Review Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_stickpro_go-store_internal_delivery_http_request_product_review_request.CreateProductReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ProductReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Create Product Review
      tags:
      - Product Review
  /api/v1/product_review/by-product/{id}:
    post:
      consumes:
      - application/json
      description: Get Product Reviews By Product ID
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        name: product_id
        type: string
      - description: '"asc" or "desc"'
        in: query
        name: sort_by_rating
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ResponseWithFullPagination-ProductReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Get Product Reviews By Product ID
      tags:
      - Product Review
...
  /v1/category/:id/products:
    get:
      consumes:
      - application/json
      description: Get paginated product variants of a category and its subcategories
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ResponseWithFullPagination-github_com_stickpro_go-store_internal_dto_EnrichedVariantDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Get Category Products
      tags:
      - Category
...
  /v1/category/{slug}/products:
    get:
      consumes:
      - application/json
      description: Get paginated, filtered and faceted product variants of a category
        and its subcategories
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      - in: query
        name: facets
        type: boolean
      - description: comma-separated UUIDs
        in: query
        name: manufacturer_id
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        name: price_max
        type: string
      - in: query
        name: price_min
        type: string
      - description: price_asc | price_desc | new | popular | name
        in: query
        name: sort
        type: string
      - description: comma-separated statuses
        in: query
        name: stock_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-VariantListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Get category products
      tags:
      - Category
...
  /v1/product/:
    get:
      consumes:
      - application/json
      description: Get products
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ResponseWithFullPagination-ProductResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Get products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Create product
      parameters:
      - description: Create product
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Create Product
      tags:
      - Product
  /v1/product/:id:
    put:
      consumes:
      - application/json
      description: Update product
      parameters:
      - description: Update product
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Update Product
      tags:
      - Product
...
  /v1/product/{slug}/:
    get:
      consumes:
      - application/json
      description: Get product by slug
      parameters:
      - description: Product Slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ProductWithMediumResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Product
      tags:
      - Product
...
  /v1/product/{slug}/reviews:
    get:
      consumes:
      - application/json
      description: Get Product Reviews By Product ID
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        name: product_id
        type: string
      - description: '"asc" or "desc"'
        in: query
        name: sort_by_rating
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ResponseWithFullPagination-ProductReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Get Product Reviews By Product ID
      tags:
      - Product Review
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"